Restructured admin feature, removed selector from pages and added them to routes. Removed crud.routes.ts, now it only lazy loads admin pages. Other changes: IGDB logo added to footer
