+++ frontend/src/app/features/admin/pages/location/location-crud.html
