+++ frontend/src/app/features/admin/pages/location/location-crud.ts
@@ -12,13 +12,12 @@ import { map, tap } from 'rxjs';
 
 import { Toaster } from '@shared/utils/toaster';
 import { Location } from '@shared/interfaces/location';
-import { LocationService } from '@features/tournament-hub/services/location.service';
+import { LocationService } from '@shared/services/location.service';
 import { LocationCrudModal } from './location-crud-modal/location-crud-modal';
 import { Pagination } from '@shared/components/pagination/pagination';
 import { SearchBar } from '@shared/components/search-bar/search-bar';
 
 @Component({
-  selector: 'admin-location-crud',
   imports: [LocationCrudModal, Pagination, SearchBar],
   templateUrl: './location-crud.html',
 })
