+++ frontend/src/app/features/tournament-hub/components/footer/footer.html
@@ -1,53 +1,43 @@
 <footer class="footer sm:footer-horizontal bg-base-300 text-neutral-content p-10">
-  <aside>
-    <h1 id="footer-title">Okizeme</h1>
-    <p>
-      Okizeme team.
-      <br />
-      Providing reliable tournament tech since 2023
-    </p>
+  <aside class="flex items-center justify-between gap-7 animate-fadeIn">
+    <div>
+      <h1 id="footer-title">Okizeme</h1>
+      <p>
+        Okizeme team.
+        <br />
+        Providing reliable tournament tech since 2023
+      </p>
+    </div>
+    <div class="h-16 shrink-0">
+      <svg
+        role="img"
+        class="h-full w-auto"
+        viewBox="0 0 24 24"
+        xmlns="http://www.w3.org/2000/svg"
+        id="Igdb--Streamline-Simple-Icons"
+      >
+        <desc>Igdb Streamline Icon: https://streamlinehq.com</desc>
+        <title>IGDB</title>
+        <path
+          d="M24 6.228c-8 0.002 -16 0 -24 0v11.543a88.875 88.875 0 0 1 2.271 -0.333 74.051 74.051 0 0 1 17.038 -0.28c1.57 0.153 3.134 0.363 4.69 0.614V6.228zm-0.706 0.707v10.013a74.747 74.747 0 0 0 -22.588 0V6.934h22.588ZM7.729 8.84a2.624 2.624 0 0 0 -1.857 0.72 2.55 2.55 0 0 0 -0.73 1.33c-0.098 0.5 -0.063 1.03 0.112 1.51 0.177 0.488 0.515 0.917 0.954 1.196 0.547 0.354 1.224 0.472 1.865 0.401a3.242 3.242 0 0 0 1.786 -0.777c-0.003 -0.724 0.002 -1.449 -0.002 -2.173 -0.725 0.004 -1.45 -0.002 -2.174 0.003 0.003 0.317 0 0.634 0.001 0.951h1.105c0.002 0.236 0 0.473 0.002 0.71 -0.268 0.196 -0.603 0.286 -0.932 0.298 -0.32 0.02 -0.65 -0.05 -0.922 -0.225a1.464 1.464 0 0 1 -0.59 -0.744c-0.18 -0.499 -0.134 -1.085 0.163 -1.53 0.23 -0.355 0.619 -0.61 1.043 -0.647a1.8 1.8 0 0 1 1.012 0.206c0.152 0.082 0.286 0.192 0.424 0.295 0.228 -0.281 0.461 -0.559 0.692 -0.838a3.033 3.033 0 0 0 -0.595 -0.403c-0.418 -0.212 -0.892 -0.285 -1.357 -0.283Zm11.66 0.086c-0.093 0 -0.187 0.002 -0.28 0 -0.68 0.002 -1.359 -0.004 -2.038 0.003 0.003 1.666 0 3.332 0.002 4.998h2.497c0.239 -0.002 0.478 -0.034 0.709 -0.097 0.276 -0.076 0.546 -0.208 0.742 -0.422 0.194 -0.208 0.297 -0.492 0.304 -0.776 0.016 -0.278 -0.032 -0.572 -0.195 -0.804 -0.175 -0.252 -0.453 -0.408 -0.734 -0.514 0.211 -0.122 0.407 -0.285 0.521 -0.505 0.134 -0.246 0.149 -0.535 0.117 -0.807a1.156 1.156 0 0 0 -0.436 -0.73c-0.264 -0.207 -0.599 -0.304 -0.93 -0.334a2.757 2.757 0 0 0 -0.279 -0.012Zm-16.715 0v5.002h1.102V8.927c-0.368 -0.002 -0.735 0 -1.102 0zm8.524 0v5.002h2.016a2.87 2.87 0 0 0 1.07 -0.211 2.445 2.445 0 0 0 1.174 -0.993c0.34 -0.555 0.429 -1.244 0.292 -1.876a2.367 2.367 0 0 0 -0.828 -1.338c-0.478 -0.387 -1.096 -0.577 -1.707 -0.584h-2.017zm6.949 0.967c0.392 0.002 0.784 -0.001 1.176 0.002 0.183 0.011 0.38 0.054 0.51 0.19 0.11 0.112 0.136 0.28 0.112 0.43a0.436 0.436 0 0 1 -0.22 0.316 1.082 1.082 0 0 1 -0.483 0.116c-0.365 0.002 -0.73 -0.001 -1.094 0.001 -0.002 -0.351 0 -0.703 -0.001 -1.054zm-5.031 0.026c0.28 0 0.567 0.053 0.815 0.19 0.274 0.149 0.491 0.396 0.607 0.685 0.113 0.272 0.138 0.574 0.107 0.865a1.456 1.456 0 0 1 -0.335 0.786 1.425 1.425 0 0 1 -0.865 0.466c-0.168 0.031 -0.34 0.022 -0.51 0.023h-0.632V9.92h0.813zm5.03 1.948h1.36c0.174 0.006 0.354 0.035 0.505 0.127 0.11 0.066 0.191 0.18 0.212 0.308 0.025 0.15 0.004 0.32 -0.099 0.44 -0.102 0.12 -0.258 0.176 -0.409 0.2 -0.172 0.032 -0.348 0.02 -0.522 0.022 -0.35 -0.001 -0.698 0.002 -1.047 -0.001v-1.096z"
+          fill="#FFFFFF"
+          stroke-width="1"
+        ></path>
+      </svg>
+    </div>
   </aside>
-  <nav>
+
+  <nav class="animate-fadeIn">
     <h6 class="footer-title">Social</h6>
     <div class="grid grid-flow-col gap-4">
-      <a>
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          width="24"
-          height="24"
-          viewBox="0 0 24 24"
-          class="fill-current"
-        >
-          <path
-            d="M24 4.557c-.883.392-1.832.656-2.828.775 1.017-.609 1.798-1.574 2.165-2.724-.951.564-2.005.974-3.127 1.195-.897-.957-2.178-1.555-3.594-1.555-3.179 0-5.515 2.966-4.797 6.045-4.091-.205-7.719-2.165-10.148-5.144-1.29 2.213-.669 5.108 1.523 6.574-.806-.026-1.566-.247-2.229-.616-.054 2.281 1.581 4.415 3.949 4.89-.693.188-1.452.232-2.224.084.626 1.956 2.444 3.379 4.6 3.419-2.07 1.623-4.678 2.348-7.29 2.04 2.179 1.397 4.768 2.212 7.548 2.212 9.142 0 14.307-7.721 13.995-14.646.962-.695 1.797-1.562 2.457-2.549z"
-          ></path>
-        </svg>
+      <a href="https://www.twitter.com" target="_blank">
+        <span class="icon-[line-md--twitter-x-alt]" style="width: 24px; height: 24px"></span>
       </a>
-      <a>
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          width="24"
-          height="24"
-          viewBox="0 0 24 24"
-          class="fill-current"
-        >
-          <path
-            d="M19.615 3.184c-3.604-.246-11.631-.245-15.23 0-3.897.266-4.356 2.62-4.385 8.816.029 6.185.484 8.549 4.385 8.816 3.6.245 11.626.246 15.23 0 3.897-.266 4.356-2.62 4.385-8.816-.029-6.185-.484-8.549-4.385-8.816zm-10.615 12.816v-8l8 3.993-8 4.007z"
-          ></path>
-        </svg>
+      <a href="https://www.youtube.com" target="_blank">
+        <span class="icon-[line-md--youtube-filled]" style="width: 24px; height: 24px"></span>
       </a>
-      <a>
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          width="24"
-          height="24"
-          viewBox="0 0 24 24"
-          class="fill-current"
-        >
-          <path
-            d="M9 8h-3v4h3v12h5v-12h3.642l.358-4h-4v-1.667c0-.955.192-1.333 1.115-1.333h2.885v-5h-3.808c-3.596 0-5.192 1.583-5.192 4.615v3.385z"
-          ></path>
-        </svg>
+      <a href="https://www.facebook.com" target="_blank">
+        <span class="icon-[line-md--facebook]" style="width: 24px; height: 24px"></span>
       </a>
     </div>
   </nav>
