+++ frontend/src/app/shared/services/game.service.ts
