+++ frontend/src/app/shared/services/tournament.service.ts
